NeutralLandscapes.jl

A pure Julia port of https://github.com/tretherington/nlmpy

Landscape models

NeutralLandscapes.NeutralLandscapeMaker — Type

All algorithms are descended from the NeutralLandscapeMaker type. A new algorithm must minimally implement and _landscape! method for this type.

NeutralLandscapes.NoGradient — Type

NoGradient

This type is used to generate a random landscape with no gradients

NeutralLandscapes.PlanarGradient — Type

PlanarGradient

This type is used to generate a planar gradient landscape, where values change as a bilinear function of the x and y coordinates. The direction is expressed as a floating point value, which will be in [0,360]. The inner constructor takes the mod of the value passed and 360, so that a value that is out of the correct interval will be corrected.

NeutralLandscapes.EdgeGradient — Type

EdgeGradient

This type is used to generate an edge gradient landscape, where values change as a bilinear function of the x and y coordinates. The direction is expressed as a floating point value, which will be in [0,360]. The inner constructor takes the mod of the value passed and 360, so that a value that is out of the correct interval will be corrected.

NeutralLandscapes.WaveSurface — Type

WaveSurface

Creates a sinusoidal landscape with a direction and a number of periods. If neither are specified, there will be a single period of random direction.

NeutralLandscapes.DistanceGradient — Type

DistanceGradient

The sources field is the linear indices of the matrix, from which the distance must be calculated.

NeutralLandscapes.RectangularCluster — Type

RectangularCluster

Fills the landscape with rectangles containing a random value. The size of each rectangle/patch is between minimum and maximum (the two can be equal for a fixed size rectangle).

NeutralLandscapes.NearestNeighborElement — Type

NearestNeighborElement

Assigns a value to each patch using a k-NN algorithmm with n initial clusters and k neighbors. The default is to use three cluster and a single neighbor.

NeutralLandscapes.PerlinNoise — Type

PerlinNoise <: NeutralLandscapeMaker

PerlinNoise(; kw...)

Create a Perlin noise neutral landscape model with values ranging 0-1.

Keywords

periods::Tuple{Int,Int}=(1,1): the number of periods of Perlin noise across row and column dimensions for the first octave.

octaves::Int=1: the number of octaves that will form the Perlin noise.

lacunarity::Int=2 : the rate at which the frequency of periods increases for each octive.

persistance::Real=0.5 : the rate at which the amplitude of periods decreases for each octive.

valley::Symbol=:u: the kind of valley bottom that will be mimicked::uproduces u-shaped valleys,:vproduces v-shaped valleys, and:-` produces flat bottomed valleys.

Note: This is a memory-intensive algorithm with some settings. Be careful using larger prime numbers for period when also using a large array size, high lacuarity and/or many octaves. Memory use scales with the lowest common multiple of periods.

Landscape generating function

Random.rand! — Function

rand!(mat, alg) where {IT <: Integer}

Fill the matrix mat with a landscape created following the model defined by alg. The mask argument accepts a matrix of boolean values, and is passed to mask! if it is not nothing.

Base.rand — Function

rand(alg, dims::Tuple{Vararg{Int64,2}}; mask=nothing) where {T <: Integer}

Creates a landscape of size dims (a tuple of two integers) following the model defined by alg. The mask argument accepts a matrix of boolean values, and is passed to mask! if it is not nothing.

Missing docstring.

Missing docstring for _landscape!. Check Documenter's build log for details.

Other functions

Missing docstring.

Missing docstring for mask!. Check Documenter's build log for details.